feat: add getToken function
Up to this point, you’ve 1) checked for a token, and 2) checked if that token is still valid. If it is, you call the Google Calendar API and get the events. However, when the token doesn’t exist or is invalid, you need to get a new one. To that end, you need to redirect your users to log in with Google so they can be redirected back to your site with the code. The new token will be fetched in the code above using the function getToken (not defined yet) , which is called if the authorization code is present.

This function takes your code and encodes it using encodeURIComponent, then uses the encoded code to get your token.