first commit: add code given from console with steps Given, When, Then”
Cucumber will run the test, notice the missing code, then provide that code for you based on the format of your Gherkin file.
Each test in the code above represents a scenario from your Gherkin file (“filterEventsByCity.feature”). You’ll notice that each one includes a list of steps (following the “Given, When, Then” syntax) that match the steps in the file.